Implement an inlining pass for ANF
It's not used at the moment, but it could be run before the
ANF->MCode translation to do some optimizations (which could be
significant for things like arithmetic code). It should _not_ be run
on the ANF terms that will be stored and exchanged, because that
will change their hashes and actually generate invalid code for
e.g. the JIT. It should only be run as a pass before generating the
final interpreter code.