Add verification for non-test transcripts
Checks that other transcripts in the source tree (currently just the GitHub
bug-report template) are valid.

This also fixes a bug where a transcript block like ` ``` ucm :hidec`
would be parsed like ` ``` ucm :hide c`, rather than complaining that
there is no `hidec` tag.

This currently fails because of a typo introduced earlier in this PR.