Add support for “unparsed” UCM command args
Previously, it was impossible to, say, pass numbers to `run foo`,
because they would be parsed as numbered args, and then would error when
a non-string was given as a command-line arg. This now replaces all
usage of `unsupportedStructuredArgument` (which would fail on numbers)
with handling of the original string.

This improves the following commands:
- `load`
- `display.to`
- `debug.tab-complete`
- `debug.lsp-name-completion`
- `debug.fuzzy-options`
- `help-topics`
- `help`
- `docs.to-html`
- `run`
- `compile`
- `run.native`
- `compile.native`
- `create.author`
- `release.draft`

Fixes #2805.