Update nix flake
This still keeps nixpkgs on the 24.05 release, because 24.11 updates stack to version 3, which isn't currently being
used by ucm devs/CI. But it updates the dependencies in the flake.lock.

This does update hpack from 0.35.2 to 0.36.1. I think that this is probably fine, because I see checked-in files that
report that they were generated by hpack 0.36. But I also have no idea what I'm doing in the Haskell world.