Remove `HQSplit` and `HQSplitAbsolute`
All uses of `HashQualified` should now be outside the full pathy
structure. There are a couple motivations for this:

1. it makes the many `Name` ↔︎ `Split` conversions simpler and
2. changing the `Path` component of a `Split` should have to pass
   through the `HashQualified` structure, because it can potentially
   invalidate the hash[^1].

[^1]: `HashQualified` shouldn’t have a `Functor` instance, because it
should only be modifiable in carefully-controlled ways, but that’s a
battle for another day.

Types like `Split` and `AbsSplit` still exist because there aren’t
`Name` equivalents to those types (and there are additional types like
`(ProjectPath, NameSegment)` that are structured similarly to `Split`).