Make `Path` operations more consistent
This adds a type class, `Pathy`, with instances for the common path
types: `Path`, `Path'`, `Absolute`, and `Relative`. The intent is to
simplify the code, but also to make it easier to transition away from
direct `Path` usage in future.

This also renames a few operations, to make them more path-specific.
E.g., instead of `snoc`, it’s `descend`[^1]. and `absoluteEmpty` is now
`root`.

[^1]: `snoc` and `cons` are particularly confusing names, because we
store segments in different orders in different contexts, and it’s not
clear whether `snoc`/`cons` should refer to the internal ordering or the
conceptual ordering. Also, `Path` is likely to reverse it order at some
point to match `Name`, and if `snoc`/`cons` reflects the internal
ordering, then that’ll break things.

Also the `Path` constructor/field are no longer exported, and `Path'` is
now defined directly in terms of the `pattern`s that were used
everywhere.