Change ForeignConvention structure
The new conventions are not aimed to encode structure on the stack,
but to instead directly produce corresponding unison values. This
should allow the raw foreign operations to be called directly in
most cases, without the wrapper having to mediate between the
convention encoding and the unison value. This both allows for more
efficient direct calls, and avoiding what might be considerable
overhead of building up the unison values via multiple iterations
through the interpreter loop.

The one portion where a 'flattened' encoding is retained is for
tuples, but this is just for supporting multi-argument functions,
not for actually encoding tuples on the stack.