Modify foreign declarations and machine implementation for new conventions
Most wrappers became obselete for foreigns, and we can just specify
how many arguments they take.

The foreign functions now have an option for indicating an
exceptional result, in which case the returned value is a `Failure`,
and the machine will directly call the `Exception` handler using that
value. This allows e.g. array operations to have less interpreter
overhead.