Various improvements to EasyTest expectations
I had a confusing failure that was made more confusing by the fact that
the reported failure had reversed the expected and actual values. This
addresses that, as well as some related testing concerns.

- swap many (if not all) misordered expected/actual arguments
- try to format them in a way that makes it clear that the expression
  (the “actual” value) is the final argument
- replace some `expect $ a == b` with `expectEqual b a` for better
  failure messages
- define `expectOrdinal` locally to simplify a group of tests
- move and generalize `expectRight'` for better error reporting of
  `Either`

The nicest bit of this change is taking advantage of some of EasyTest’s
compositionality. Previously, there was code like `expectEqual (Just x)
. either (const Nothing) Just`, which would discard a `Left` that lacked
an `Eq` instance; or `expectRight` which never checked for the correct
result. Now `expectEqual x <=< expectRight'` doesn’t require an `Eq` on
the `Left` , but will `show` the `Left` if there’s a failure. And
removing the final `'` will remove the need for `Show`, while still
allowing the comparison of the `Right`.