Rework data pattern matching to use default cases
Previously, we were always expanding data type matching to explicitly
match against every case. This was necessary because we would eagerly
dump the entire data contents to the stack and analyze the tag from
there. Different constructors with different numbrs of arguments would
result in a different stack layout, so default cases could not be
uniform.

However, a while ago we started analyzing data type tags directly, to
avoid putting them on the stack. This change goes the final step. It
selects the branch before deciding to dump the data type to the stack,
and only does so in specific case branches. Nothing is added to the stack
in default branches, so they can be shared among all constructors.

On the other end, the pattern matching compiler has been changed to
preserve default cases when possible. When splitting on a data type,
the actual constructors matched in the source are used as the explicit
branches, and a default case is used if it exists in the source. The
translation will still sometimes duplicate branches, but not nearly as
much as before (and mainly due to complicated pattern matching).