Show the names of tests that fail sandbox violation
Previously if a test violated the sandbox such as with `Debug.trace`, the output
just said `pure code can't perform I/O` but didn't actually report any failing
tests or even name the test that violated the sandbox.

With this change it will at least add a line like:

```
Error while evaluating test `bar.test`
```

This should help a lot in tracking down the sandbox violation.

This resolves the worst of [#4685](https://github.com/unisonweb/unison/issues/4685), but it doesn't do anything to address the confusing behavior of test watches being exempt from the sandbox and caching tests that would fail when running `test` on an empty cache.