Modify the fix-2805 transcript to use return type
It previously used `printLine` to print the result from the `main`
function, but transcripts don’t capture stdout, so this changes it to
return the result instead, so it’s reflected in the output.

This is a follow-up to #5549, based on [this comment](https://github.com/unisonweb/unison/pull/5549#discussion_r1947962933).