Fix a problem with type lookups in certain kind checking scenarios
In the `run` command, when running something from the latest
checked file, any data and effect decls would remain in the
UnisonFile used to run the expression, but their type
dependencies would not be fetched from the codebase. Because
they're in the file, they must be kind checked, but that would
only work if the main expression depends on the same types as the
data declarations.